+++ src/GitHub/delete.ts
